Itch Changes

+++ src/containers/sprite-selector-item.jsx
@@ -98,7 +98,9 @@ class SpriteSelectorItem extends React.Component {
     }
     handleDelete (e) {
         e.stopPropagation(); // To prevent from bubbling back to handleClick
-        this.props.onDeleteButtonClick(this.props.id);
+        const confirmToDelete = confirm('Are you sure that you want to delete this Sprite?');
+        if(confirmToDelete)
+            this.props.onDeleteButtonClick(this.props.id);
     }
     handleDuplicate (e) {
         e.stopPropagation(); // To prevent from bubbling back to handleClick
